# web-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blob Championship System API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/blobs/all": {
      "get": {
        "tags": [
          "blobs"
        ],
        "summary": "Get All",
        "operationId": "get_all_blobs_all_get",
        "parameters": [
          {
            "name": "name_search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name Search"
            }
          },
          {
            "name": "show_dead",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Show Dead"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlobStatsDto"
                  },
                  "title": "Response Get All Blobs All Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/blobs/create": {
      "post": {
        "tags": [
          "blobs"
        ],
        "summary": "Create Blob",
        "operationId": "create_blob_blobs_create_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "first_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "First Name"
            }
          },
          {
            "name": "last_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Last Name"
            }
          },
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/factory/save-name-suggestion": {
      "post": {
        "tags": [
          "factory"
        ],
        "summary": "Save Name Suggestion",
        "operationId": "save_name_suggestion_factory_save_name_suggestion_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "first_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "First Name"
            }
          },
          {
            "name": "last_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Last Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/factory/update-name-suggestion": {
      "post": {
        "tags": [
          "factory"
        ],
        "summary": "Update Name Suggestion",
        "description": "Updates the first name of a child name suggestion",
        "operationId": "update_name_suggestion_factory_update_name_suggestion_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          },
          {
            "name": "first_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "First Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/competition/quartered-event-results": {
      "post": {
        "tags": [
          "competition"
        ],
        "summary": "Save Quartered",
        "description": "Save event results for quartered events.",
        "operationId": "save_quartered_competition_quartered_event_results_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_save_quartered_competition_quartered_event_results_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/competition/race-event-results": {
      "post": {
        "tags": [
          "competition"
        ],
        "summary": "Save Race",
        "description": "Save event results for race events.",
        "operationId": "save_race_competition_race_event_results_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_save_race_competition_race_event_results_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/actions/create/quartered": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Quartered",
        "description": "Generate score for contender and save the action for given event.",
        "operationId": "quartered_actions_create_quartered_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "tick",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tick"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlobCompetitorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/actions/create/race": {
      "post": {
        "tags": [
          "actions"
        ],
        "summary": "Race",
        "description": "Generate score for all contenders and save the actions for given event.",
        "operationId": "race_actions_create_race_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          },
          {
            "name": "tick",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tick"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/validate": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Validate Token",
        "operationId": "validate_token_auth_validate_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ActionDto": {
        "properties": {
          "tick": {
            "type": "integer",
            "title": "Tick"
          },
          "blob_id": {
            "type": "integer",
            "title": "Blob Id"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "tick",
          "blob_id",
          "score"
        ],
        "title": "ActionDto"
      },
      "BlobCompetitorDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "strength": {
            "type": "number",
            "title": "Strength"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "strength",
          "color"
        ],
        "title": "BlobCompetitorDto"
      },
      "BlobStatsDto": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "born": {
            "type": "string",
            "title": "Born"
          },
          "debut": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debut"
          },
          "contract": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contract"
          },
          "podiums": {
            "type": "integer",
            "title": "Podiums"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "championships": {
            "type": "integer",
            "title": "Championships"
          },
          "grandmasters": {
            "type": "integer",
            "title": "Grandmasters"
          },
          "league_name": {
            "type": "string",
            "title": "League Name"
          },
          "at_risk": {
            "type": "boolean",
            "title": "At Risk",
            "default": false
          },
          "is_dead": {
            "type": "boolean",
            "title": "Is Dead",
            "default": false
          },
          "is_retired": {
            "type": "boolean",
            "title": "Is Retired",
            "default": false
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "name",
          "born",
          "podiums",
          "wins",
          "championships",
          "grandmasters",
          "league_name",
          "color"
        ],
        "title": "BlobStatsDto"
      },
      "Body_save_quartered_competition_quartered_event_results_post": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventDto"
          },
          "event_records": {
            "items": {
              "$ref": "#/components/schemas/QuarteredEventRecordDto"
            },
            "type": "array",
            "title": "Event Records"
          }
        },
        "type": "object",
        "required": [
          "event",
          "event_records"
        ],
        "title": "Body_save_quartered_competition_quartered_event_results_post"
      },
      "Body_save_race_competition_race_event_results_post": {
        "properties": {
          "event": {
            "$ref": "#/components/schemas/EventDto"
          },
          "event_records": {
            "items": {
              "$ref": "#/components/schemas/RaceEventRecordDto"
            },
            "type": "array",
            "title": "Event Records"
          }
        },
        "type": "object",
        "required": [
          "event",
          "event_records"
        ],
        "title": "Body_save_race_competition_race_event_results_post"
      },
...
      "EventDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "competitors": {
            "items": {
              "$ref": "#/components/schemas/BlobCompetitorDto"
            },
            "type": "array",
            "title": "Competitors"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionDto"
            },
            "type": "array",
            "title": "Actions"
          },
          "league": {
            "$ref": "#/components/schemas/LeagueDto"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          }
        },
        "type": "object",
        "required": [
          "id",
          "competitors",
          "actions",
          "league",
          "season",
          "round",
          "type"
        ],
        "title": "EventDto"
      },
...
      "EventType": {
        "type": "string",
        "enum": [
          "QUARTERED_ONE_SHOT_SCORING",
          "QUARTERED_TWO_SHOT_SCORING",
          "ENDURANCE_RACE"
        ],
        "title": "EventType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LeagueDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "field_size": {
            "type": "integer",
            "title": "Field Size"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "field_size",
          "level"
        ],
        "title": "LeagueDto"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "QuarteredEventRecordDto": {
        "properties": {
          "blob": {
            "$ref": "#/components/schemas/BlobCompetitorDto"
          },
          "quarters": {
            "items": {
              "$ref": "#/components/schemas/ScoreDto"
            },
            "type": "array",
            "title": "Quarters"
          },
          "eliminated": {
            "type": "boolean",
            "title": "Eliminated",
            "default": false
          },
          "current": {
            "type": "boolean",
            "title": "Current",
            "default": false
          },
          "next": {
            "type": "boolean",
            "title": "Next",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "blob",
          "quarters"
        ],
        "title": "QuarteredEventRecordDto"
      },
      "RaceEventRecordDto": {
        "properties": {
          "blob": {
            "$ref": "#/components/schemas/BlobCompetitorDto"
          },
          "distance_records": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Distance Records"
          },
          "previous_position": {
            "type": "integer",
            "title": "Previous Position",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "blob",
          "distance_records"
        ],
        "title": "RaceEventRecordDto"
      },
      "ScoreDto": {
        "properties": {
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "best": {
            "type": "boolean",
            "title": "Best",
            "default": false
          },
          "personal_best": {
            "type": "boolean",
            "title": "Personal Best",
            "default": false
          },
          "latest_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Score"
          }
        },
        "type": "object",
        "title": "ScoreDto"
      },
...
      "Token": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "Token"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        }
      }
    }
  }
}